Unused project idea/requirements.txt
pandas == 2.2.3
scikit-learn == 1.5.2
joblib == 1.4.2
django == 5.1.2
djangorestframework == 3.15.2
django-cors-headers == 4.5.0